refactor of carp_firebase based on testing
