+++ carp_mobile_sensing_app/test/widget_test.dart
@@ -8,7 +8,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:carp_mobile_sensing_app/old_main.dart';
+import 'package:carp_mobile_sensing_app/flutter_demo.dart';
 
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
